refactor(mode): move tmux UI into src/ui/tmux
