+++ src/action/tmux/rename_session.rs
@@ -1,7 +1,7 @@
 //! `RenameSessionRequest` — opens the ozmux rename prompt pre-filled with the
 //! target session's current name.
 
-use crate::mode::tmux::rename_prompt::{RenamePrompt, RenameSubject};
+use crate::ui::tmux::rename_prompt::{RenamePrompt, RenameSubject};
 use bevy::prelude::*;
 use ozmux_tmux::TmuxSession;
