+++ src/mode/tmux.rs
@@ -1,24 +1,11 @@
 //! tmux feature plugin: aggregates all tmux runtime sub-plugins.
 
-pub(crate) mod confirm_prompt;
-mod divider_handle;
-pub(crate) mod mode_ui;
-mod pane_focus;
-pub(crate) mod rename_prompt;
-pub(crate) mod window_bar;
-
 use crate::input::tmux::forward::ForwardPlugin;
 use crate::input::tmux::gate::GatePlugin;
 use crate::input::tmux::input::InputPlugin;
 use crate::input::tmux::mouse::MousePlugin;
 use crate::input::tmux::window_bar_input::WindowBarInputPlugin;
 use bevy::prelude::*;
-use confirm_prompt::ConfirmPromptPlugin;
-use divider_handle::DividerHandlePlugin;
-use mode_ui::TmuxModeUiPlugin;
-use pane_focus::PaneFocusPlugin;
-use rename_prompt::RenamePromptPlugin;
-use window_bar::WindowBarPlugin;
 
 /// Bevy plugin aggregating all tmux runtime sub-plugins.
 pub struct OzmuxTmuxPlugin;
@@ -29,12 +16,8 @@ impl Plugin for OzmuxTmuxPlugin {
             InputPlugin,
             MousePlugin,
             ForwardPlugin,
-            WindowBarPlugin,
             WindowBarInputPlugin,
-            DividerHandlePlugin,
-            PaneFocusPlugin,
             GatePlugin,
-        ))
-        .add_plugins((TmuxModeUiPlugin, ConfirmPromptPlugin, RenamePromptPlugin));
+        ));
     }
 }
