+++ src/render/tmux.rs
@@ -7,10 +7,10 @@ pub(crate) mod copy_mode;
 mod paint_rescue;
 
 use crate::app_mode::TmuxActiveSet;
-use crate::mode::tmux::mode_ui::WorkspaceUiRoot;
 use crate::surface::OzmaTerminal;
 use crate::surface_geom::cells_for;
 use crate::theme;
+use crate::ui::tmux::mode_ui::WorkspaceUiRoot;
 use bevy::ecs::message::MessageReader;
 use bevy::math::Rect;
 use bevy::prelude::*;
