+++ src/ui/tmux/window_bar.rs
